+++ CHANGES.md
@@ -1,4 +1,5 @@
 Release with following new features and fixes:
+- Add support for reMarkable Paper Pro
 - Build against .NET 8.0.11
 - Build against Avalonia 11.2.2
 - Build against SSH.NET 2024.2.0