+++ CHANGES.md
@@ -1,3 +1,3 @@
 Release with following new features and fixes:
 - Build against Avalonia TreeDataGrid 11.1.1
-- Build against Microsoft Entity Framework Core 9.0.1
+- Build against Microsoft Entity Framework Core 8.0.12