+++ CHANGES.md
@@ -1,3 +1,4 @@
 Release with following new features and fixes:
 - Build against Avalonia TreeDataGrid 11.1.1
 - Build against Microsoft Entity Framework Core 8.0.12
+- Build against NLog 5.4.0